+++ internal/konnect/node_agent.go
@@ -1,15 +1,22 @@
 package konnect
 
 import (
+	"context"
 	"fmt"
+	"sync/atomic"
 	"time"
 
 	"github.com/go-logr/logr"
 
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
 )
 
-const defaultRefreshNodeInterval = 30 * time.Second
+const (
+	MinRefreshNodePeriod     = 30 * time.Second
+	DefaultRefreshNodePeriod = 60 * time.Second
+	NodeOutdateInterval      = 5 * time.Minute
+)
 
 type NodeAgent struct {
 	NodeID   string
@@ -18,24 +25,44 @@ type NodeAgent struct {
 
 	Logger logr.Logger
 
-	konnectClient   *NodeAPIClient
-	refreshInterval time.Duration
+	konnectClient *NodeAPIClient
+	refreshPeriod time.Duration
+
+	configStatus           atomic.Uint32
+	configStatusSubscriber dataplane.ConfigStatusSubscriber
 }
 
-func NewNodeAgent(hostname string, version string, logger logr.Logger, client *NodeAPIClient) *NodeAgent {
-	return &NodeAgent{
+func NewNodeAgent(
+	hostname string,
+	version string,
+	refreshPeriod time.Duration,
+	logger logr.Logger,
+	client *NodeAPIClient,
+	configStatusSubscriber dataplane.ConfigStatusSubscriber,
+) *NodeAgent {
+	if refreshPeriod < MinRefreshNodePeriod {
+		refreshPeriod = MinRefreshNodePeriod
+	}
+	a := &NodeAgent{
 		Hostname: hostname,
 		Version:  version,
 		Logger: logger.
 			WithName("konnect-node").WithValues("runtime_group_id", client.RuntimeGroupID),
-		konnectClient: client,
-		// TODO: set refresh interval by some flag
-		// https://github.com/Kong/kubernetes-ingress-controller/issues/3515
-		refreshInterval: defaultRefreshNodeInterval,
+		konnectClient:          client,
+		refreshPeriod:          refreshPeriod,
+		configStatusSubscriber: configStatusSubscriber,
 	}
+	a.configStatus.Store(uint32(dataplane.ConfigStatusOK))
+	return a
 }
 
 func (a *NodeAgent) createNode() error {
+	err := a.clearOutdatedNodes()
+	if err != nil {
+		// still continue to update the current status if cleanup failed.
+		a.Logger.Error(err, "failed to clear outdated nodes")
+	}
+
 	createNodeReq := &CreateNodeRequest{
 		ID:       a.NodeID,
 		Hostname: a.Hostname,
@@ -45,7 +72,7 @@ func (a *NodeAgent) createNode() error {
 	}
 	resp, err := a.konnectClient.CreateNode(createNodeReq)
 	if err != nil {
-		return fmt.Errorf("failed to update node, hostname %s: %w", a.Hostname, err)
+		return fmt.Errorf("failed to create node, hostname %s: %w", a.Hostname, err)
 	}
 
 	a.NodeID = resp.Item.ID
@@ -60,7 +87,16 @@ func (a *NodeAgent) clearOutdatedNodes() error {
 	}
 
 	for _, node := range nodes.Items {
-		if node.Type == NodeTypeIngressController && node.Hostname != a.Hostname {
+		deleteNode := false
+		if node.Type == NodeTypeIngressController {
+			// nodes to remove:
+			// (1) since only one KIC node is allowed in a runtime group, all the nodes with other hostnames are considered outdated.
+			// (2) in some cases(kind/minikube restart), rebuilt pod uses the same name. So nodes updated for >5mins before should be deleted.
+			if node.Hostname != a.Hostname || time.Since(time.Unix(node.UpdatedAt, 0)) > NodeOutdateInterval {
+				deleteNode = true
+			}
+		}
+		if deleteNode {
 			a.Logger.V(util.DebugLevel).Info("remove outdated KIC node", "node_id", node.ID, "hostname", node.Hostname)
 			err := a.konnectClient.DeleteNode(node.ID)
 			if err != nil {
@@ -71,16 +107,40 @@ func (a *NodeAgent) clearOutdatedNodes() error {
 	return nil
 }
 
+func (a *NodeAgent) subscribeConfigStatus(ctx context.Context) {
+	ch := a.configStatusSubscriber.SubscribeConfigStatus()
+	chDone := ctx.Done()
+
+	for {
+		select {
+		case <-chDone:
+			a.Logger.Info("subscribe loop stopped", "message", ctx.Err().Error())
+			return
+		case configStatus := <-ch:
+			a.configStatus.Store(uint32(configStatus))
+		}
+	}
+}
+
 func (a *NodeAgent) updateNode() error {
 	err := a.clearOutdatedNodes()
 	if err != nil {
+		// still continue to update the current status if cleanup failed.
 		a.Logger.Error(err, "failed to clear outdated nodes")
-		return err
 	}
 
-	// TODO: retrieve the real state of KIC
-	// https://github.com/Kong/kubernetes-ingress-controller/issues/3515
-	ingressControllerStatus := IngressControllerStateOperational
+	var ingressControllerStatus IngressControllerState
+	configStatus := int(a.configStatus.Load())
+	switch dataplane.ConfigStatus(configStatus) {
+	case dataplane.ConfigStatusOK:
+		ingressControllerStatus = IngressControllerStateOperational
+	case dataplane.ConfigStatusTranslationErrorHappened:
+		ingressControllerStatus = IngressControllerStatePartialConfigFail
+	case dataplane.ConfigStatusApplyFailed:
+		ingressControllerStatus = IngressControllerStateInoperable
+	default:
+		ingressControllerStatus = IngressControllerStateUnknown
+	}
 
 	updateNodeReq := &UpdateNodeRequest{
 		Hostname: a.Hostname,
@@ -99,24 +159,30 @@ func (a *NodeAgent) updateNode() error {
 	return nil
 }
 
-func (a *NodeAgent) updateNodeLoop() {
-	ticker := time.NewTicker(a.refreshInterval)
+func (a *NodeAgent) updateNodeLoop(ctx context.Context) {
+	ticker := time.NewTicker(a.refreshPeriod)
 	defer ticker.Stop()
-	// TODO: add some mechanism to break the loop
-	// https://github.com/Kong/kubernetes-ingress-controller/issues/3515
-	for range ticker.C {
-		err := a.updateNode()
-		if err != nil {
-			a.Logger.Error(err, "failed to update node", "node_id", a.NodeID)
+	for {
+		select {
+		case <-ctx.Done():
+			err := ctx.Err()
+			a.Logger.Info("update node loop stopped", "message", err.Error())
+			return
+		case <-ticker.C:
+			err := a.updateNode()
+			if err != nil {
+				a.Logger.Error(err, "failed to update node", "node_id", a.NodeID)
+			}
 		}
 	}
 }
 
-func (a *NodeAgent) Run() {
+func (a *NodeAgent) Run(ctx context.Context) {
 	err := a.createNode()
 	if err != nil {
 		a.Logger.Error(err, "failed to create node, agent abort")
 		return
 	}
-	go a.updateNodeLoop()
+	go a.updateNodeLoop(ctx)
+	go a.subscribeConfigStatus(ctx)
 }